requirements.txt
torch>=2.2.0
transformers>=4.44.2
accelerate>=0.33.0
safetensors>=0.4.4
Pillow>=10.3.0
sentencepiece>=0.2.0
qwen-vl-utils>=0.0.8
requests>=2.31.0
numpy>=1.26.4
torchvision>=0.20.0
